Finally, something working. Fabric drawing.
At long last, a freaking updating drawing. Fabicjs is powering this, and
it was ridiculously easy to set up. There's lots of other code to
change, but now I can condense the nonsense and start building an actual
project. Will move all the physics logic into it's own js file, and then
all the rendering in another file. The structure will be similar to
debug draw but built on fabricjs. Then I'll import XML files to test
Sodarace/MINS objects.