Load XML complete. Works mostly.
There are a lot of parameters coming from this loading process that just
don't translate to the new system. I'm inclined to ignore those incoming
settings, and just put in more reasonable settings. Either way, good
progress here, as there are routes for creating joints, and new bodies,
as well as callbacks to accompany them.

Loading creates objects with the original shapes in tact. However, the
muscles just don't work. Not sure on the next step. Maybe loading
terrain, maybe not. The hardest remaining part I believe is wiring this
to the C# code and defining bodies created by CPPNs (that's done C# side
as well).