Infinite scroll working, offscreen objects disabled.
Large memory footprint for long scrolling, but it definitely works. I
would like to redesign this in some way to be more effective and paginate
in some meaningful way.

Being able to scroll forever isn't as desirable as originally thought.
Possible, but not very useful. Now, I'm also thinking I'd like to have a
"pinterest" style front page for displaying all the items found with WIN
in the domain. That's for much later.